+++ internal/server/repository/postgres/test/test_utils.go
@@ -51,7 +51,7 @@ func createTable(ctx context.Context, conn pgx.Conn, tableName string) error {
 		(
 			id    varchar not null,
 			type  varchar not null,
-			delta integer,
+			delta int8,
 			value double precision,
 			constraint #T_pk
 				unique (id, type)
